rx_detect: int32 magnitudes, and the arena reaches its balance point
The ZC mag ring was int64 holding values whose measured peak is 780982
-- 20 bits, a 2750x margin -- so int32 halves it: 61448 -> 30724. The cc
sums that gather 16 of these stay int64, as does the squared metric.

Arena 156680 -> 131584. Station RAM 723 -> 699 KB without EXT frames,
954 -> 929 KB with.

This is where the arena stops, and the reason is structural rather than
a lack of ideas: its three phases are now within 6 KB of each other
(detect 125956, demod 131584, decode 131072). A union costs the MAXIMUM,
so shrinking one buys nothing until all three come down. Streaming
eval_hyp's derotation per tile is exact and would cut demod to 65792 --
and would move the arena by 512 bytes, because decode binds immediately.
Not worth touching a hot path for.

Also corrects something I had written into FEASIBILITY.md: it called the
slide window "40960-sample", which is its size in BYTES per array. It is
10240 samples, and that is already its floor -- the scan's live span IS
preamble_len+1 = 8193, and the extra 2048 is one slide block. A smaller
block trades RAM for 8x the memmove traffic, which is a poor deal
against 9% CPU headroom.

Floor is now 288 KB ring + 129 KB arena + 94 KB summaries = 511 KB
against the H723's ~496 KB usable, so the H743-class verdict stands.

Bit-exact: 11/16/13/22/8/38/4 on x86, clean under ASan+UBSan, and
test_stream 8/8 on Cortex-M7.

+++ cport/src/rx_detect.c
@@ -29,7 +29,7 @@
 #define DET_OFF_WI   0
 #define DET_OFF_WQ   (DET_OFF_WI + ZC_SLIDE_W * (int)sizeof(samp_t))
 #define DET_OFF_MAG  (DET_OFF_WQ + ZC_SLIDE_W * (int)sizeof(samp_t))
-#define DET_OFF_KR   (DET_OFF_MAG + ZC_MAG_RING * (int)sizeof(int64_t))
+#define DET_OFF_KR   (DET_OFF_MAG + ZC_MAG_RING * (int)sizeof(int32_t))
 #define DET_OFF_KI   (DET_OFF_KR + ZC_KLEN_MAX * (int)sizeof(samp_t))
 #define DET_OFF_RRE  (DET_OFF_KI + ZC_KLEN_MAX * (int)sizeof(samp_t))
 #define DET_OFF_RIM  (DET_OFF_RRE + ZC_KLEN_MAX * (int)sizeof(samp_t))
@@ -435,7 +435,9 @@ static int detect_zc(const det_mode_t *d, const zc_src_t *src,
      *   mag[] genuinely needs a window, because cc[m] gathers
      *     mag[m + g*klen] for g < ng -- but that is (ng-1)*klen+1
      *     entries, not the whole span. */
-    int64_t *const mag = (int64_t *)(void *)(rx_arena + DET_OFF_MAG);
+    /* int32: measured peak magnitude is 780982 (20 bits), a 2750x
+     * margin. The cc sums that gather 16 of these stay int64. */
+    int32_t *const mag = (int32_t *)(void *)(rx_arena + DET_OFF_MAG);
     samp_t *const kr = (samp_t *)(rx_arena + DET_OFF_KR);
     samp_t *const ki = (samp_t *)(rx_arena + DET_OFF_KI);
     samp_t *const rom_re = (samp_t *)(rx_arena + DET_OFF_RRE);
@@ -491,7 +493,7 @@ static int detect_zc(const det_mode_t *d, const zc_src_t *src,
             aa = a < 0 ? -a : a;
             bb = b2 < 0 ? -b2 : b2;
             mag[p % ZC_MAG_RING] =
-                (aa > bb ? aa : bb) + ((aa < bb ? aa : bb) >> 1);
+                (int32_t)((aa > bb ? aa : bb) + ((aa < bb ? aa : bb) >> 1));
 
             /* mag[m .. m+span] are now resident, so cc[m] is complete */
             m = p - span;
